+++ dfp/value.go
@@ -1,9 +1,9 @@
 // Copyright 2020 Aleksandr Demakin. All rights reserved.
 
-// Package fixed implements a fixed-point number, where both mantissa
+// Package dfp implements a decimal floating-point number, where both mantissa
 // and exponent are stored in a single number.
 // Can be used to represent currency rates with up to 16 digits of precision.
-package fixed
+package dfp
 
 import (
 	"encoding/json"
@@ -117,7 +117,7 @@ type (
 	expType = int32
 )
 
-// Value is an unsigned fixed-point number.
+// Value is a positive floating-point number.
 // It currently uses a uint64 value as a data type, where
 // 8 bits are used for exponent and 56 for mantissa.
 //   63      55                                                     0